Another Mini Dungeon in the Forests
Ever tired of seeing the Orc Dungeon, and thinking to yourself. Huh, this doesnt feel like a generic orc fantasy dungeon. Where are the pallisades? Where are the Tents! The shitty living conditions!

Well ive got the solution for you. Adds in another mini dungeon in the forests, giving for more reasons to travel south the town.

Current Random Loot Spawns
1 Medical
2 Tool
1 Armor
1 Weapon
The rest being sewer

Not meant to be the best quality of dungeon related items. But just a mini dungeon around for people to explore.

(Will be working on more map changes to come for quality of life, and such)

Images:
https://media.discordapp.net/attachments/1163816776920993802/1387721623112646746/Screenshot_2025-06-26_180837.png?ex=685e5ff9&is=685d0e79&hm=c239cfa905f229a2200763644a026a47a8e5f67768b39cf3e547ac3cf3d42e73&=&format=webp&quality=lossless&width=783&height=800

https://media.discordapp.net/attachments/1163816776920993802/1387721623498526892/Screenshot_2025-06-26_180826.png?ex=685e5ff9&is=685d0e79&hm=1f6fc8a8020a0554234a9ee433d17a201062c79227fcbbd80ebc2b1066f61797&=&format=webp&quality=lossless&width=904&height=800